% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_names.R
\name{change_names}
\alias{change_names}
\title{Function destined to change names of the INMET dataset}
\usage{
change_names(data)
}
\arguments{
\item{data}{A INMET dataset}
}
\description{
Function destined to change names of the INMET dataset
}
