% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_inmet.R
\name{get_base_inmet}
\alias{get_base_inmet}
\title{Function destined to get climate information for all meteorological stations in Brazil}
\usage{
get_base_inmet(year)
}
\arguments{
\item{year}{Year of interest}
}
\description{
Function destined to get climate information for all meteorological stations in Brazil
}
