% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clima_info.R
\name{get_clima_info}
\alias{get_clima_info}
\title{Process INMET data}
\usage{
get_clima_info(file)
}
\arguments{
\item{file}{File containing information about one station in INMET}
}
\description{
Process INMET data
}
