% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_names.R
\name{new_names}
\alias{new_names}
\title{Function destined to change names}
\usage{
new_names(data, oldname, newname)
}
\arguments{
\item{data}{Data}

\item{oldname}{Old name}

\item{newname}{New name}
}
\description{
Function destined to change names
}
